Moved Drew's class files into the project
